// frontend/src/types/Entertainers.ts
export interface entertainer{
    entertainerId: number;
    entStageName: string;
    entSSN: string;
    entStreetAddress: string;
    entCity: string;
    entState: string;
    entZipCode: string;
    entPhoneNumber: string;
    entWebPage: string;
    entEMailAddress: string;
    dateEntered: string;
}